requirements.txt
asgiref==3.7.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
crispy-bootstrap5==2023.10
cryptography==39.0.1
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
ecdsa==0.19.0
enum-compat==0.0.3
filelock==3.14.0
gunicorn==21.2.0
hubspot-api-client==9.0.0
idna==3.4
intuit-oauth==1.2.5
numpy==1.26.4
oauthlib==3.2.2
packaging==23.1
psycopg==3.1.18
psycopg-binary==3.1.18
pyasn1==0.6.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-quickbooks==0.9.7
python3-openid==3.2.0
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.1.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.3
tldextract==5.1.2
typing_extensions==4.9.0
urllib3==1.26.14
wheel==0.43.0
whitenoise==6.6.0
